refactor: mudando services da empresa para pasta unica
