+++ src/services/empresaServices/pesquisarEmpresaService.ts
@@ -1,5 +1,5 @@
-import { Empresa } from '../models/empresaModel';
-import { pesquisarEmpresaPeloIdRepository} from '../repositories/empresaRepository';
+import { Empresa } from '../../models/empresaModel';
+import { pesquisarEmpresaPeloIdRepository} from '../../repositories/empresaRepository';
 
 export const pesquisarEmpresaService = async(IdEmpresa: number): Promise<Empresa|null> => {
   return await pesquisarEmpresaPeloIdRepository(IdEmpresa);
